requirements.txt
astroid==2.2.5
Django==2.2.4
django-money==0.15
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2-binary==2.8.3
py-moneyed==0.8.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
